# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rundinova Budget Alimentaire API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Categorie"
    post:
      operationId: createCategorie
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategorieInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categorie"
  /categories/{id}:
    put:
      operationId: updateCategorie
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategorieInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categorie"
    delete:
      operationId: deleteCategorie
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /produits:
    get:
      operationId: getProduits
      tags: [produits]
      summary: Get all products
      parameters:
        - name: actif
          in: query
          schema:
            type: boolean
        - name: categorie_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produit"
    post:
      operationId: createProduit
      tags: [produits]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProduitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
  /produits/{id}:
    put:
      operationId: updateProduit
      tags: [produits]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProduitInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
    delete:
      operationId: deleteProduit
      tags: [produits]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: Get all monthly budgets
      responses:
        "200":
          description: Budgets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetMensuel"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a monthly budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetMensuelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetMensuel"
  /budgets/{id}:
    get:
      operationId: getBudget
      tags: [budgets]
      summary: Get a monthly budget with product breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetDetail"
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a monthly budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetMensuelInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetMensuel"
  /budgets/{id}/produits:
    put:
      operationId: updateBudgetProduits
      tags: [budgets]
      summary: Update product budgets for a month
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetProduitsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    UpdateBudgetProduitsRequest:
      type: object
      properties:
        produits:
          type: array
          items:
            $ref: "#/components/schemas/BudgetProduitInput"
      required: [produits]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Categorie:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        emoji:
          type: string
        couleur:
          type: string
        ordre:
          type: integer
        actif:
          type: integer
        cree_le:
          type: string
      required: [id, nom]
    CategorieInput:
      type: object
      properties:
        nom:
          type: string
        emoji:
          type: string
        couleur:
          type: string
        ordre:
          type: integer
        actif:
          type: integer
      required: [nom]
    Produit:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        nom:
          type: string
        nom_local:
          type: string
        categorie_id:
          type: integer
        categorie_nom:
          type: string
        categorie_emoji:
          type: string
        categorie_couleur:
          type: string
        emoji:
          type: string
        unite:
          type: string
        frequence:
          type: string
        quantite_prevue_mois:
          type: number
        prix_unitaire:
          type: number
        cout_mensuel_prevu:
          type: number
        notes:
          type: string
        actif:
          type: integer
        ordre:
          type: integer
        cree_le:
          type: string
        modifie_le:
          type: string
      required: [id, code, nom, unite]
    ProduitInput:
      type: object
      properties:
        code:
          type: string
        nom:
          type: string
        nom_local:
          type: string
        categorie_id:
          type: integer
        emoji:
          type: string
        unite:
          type: string
        frequence:
          type: string
        quantite_prevue_mois:
          type: number
        prix_unitaire:
          type: number
        cout_mensuel_prevu:
          type: number
        notes:
          type: string
        actif:
          type: integer
        ordre:
          type: integer
      required: [nom, unite]
    BudgetMensuel:
      type: object
      properties:
        id:
          type: integer
        mois:
          type: string
        annee:
          type: integer
        budget_total_prevu:
          type: number
        nombre_personnes:
          type: integer
        notes:
          type: string
        statut:
          type: string
        cree_le:
          type: string
      required: [id, mois, annee, budget_total_prevu]
    BudgetMensuelInput:
      type: object
      properties:
        mois:
          type: string
        annee:
          type: integer
        budget_total_prevu:
          type: number
        nombre_personnes:
          type: integer
        notes:
          type: string
        statut:
          type: string
      required: [mois, annee, budget_total_prevu]
    BudgetProduit:
      type: object
      properties:
        id:
          type: integer
        budget_mensuel_id:
          type: integer
        produit_id:
          type: integer
        produit_nom:
          type: string
        produit_emoji:
          type: string
        categorie_nom:
          type: string
        quantite_prevue:
          type: number
        cout_prevu:
          type: number
        prix_unitaire_ref:
          type: number
        montant_depense:
          type: number
        notes:
          type: string
      required: [id, budget_mensuel_id, produit_id]
    BudgetProduitInput:
      type: object
      properties:
        produit_id:
          type: integer
        quantite_prevue:
          type: number
        cout_prevu:
          type: number
        prix_unitaire_ref:
          type: number
        notes:
          type: string
      required: [produit_id]
    BudgetDetail:
      type: object
      properties:
        budget:
          $ref: "#/components/schemas/BudgetMensuel"
        produits:
          type: array
          items:
            $ref: "#/components/schemas/BudgetProduit"
        total_depense:
          type: number
        total_prevu:
          type: number
      required: [budget, produits, total_depense, total_prevu]
...