# lib/api-zod/src/generated/types/budgetDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rundinova Budget Alimentaire API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetMensuel } from "./budgetMensuel";
import type { BudgetProduit } from "./budgetProduit";

export interface BudgetDetail {
  budget: BudgetMensuel;
  produits: BudgetProduit[];
  total_depense: number;
  total_prevu: number;
}