# lib/api-zod/src/generated/types/budgetMensuel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rundinova Budget Alimentaire API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetMensuel {
  id: number;
  mois: string;
  annee: number;
  budget_total_prevu: number;
  nombre_personnes?: number;
  notes?: string;
  statut?: string;
  cree_le?: string;
}