# lib/api-zod/src/generated/types/budgetProduit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rundinova Budget Alimentaire API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetProduit {
  id: number;
  budget_mensuel_id: number;
  produit_id: number;
  produit_nom?: string;
  produit_emoji?: string;
  categorie_nom?: string;
  quantite_prevue?: number;
  cout_prevu?: number;
  prix_unitaire_ref?: number;
  montant_depense?: number;
  notes?: string;
}